deploy-agent/requirements.txt
-e .

# tests
tox==1.6.1
coverage
pytest==2.3.5
mock
flake8
nose-socket-whitelist
